--- packages/express-cargo/src/validator.ts
@@ -1,953 +1,0 @@
-import {
-    AppliedDecorator,
-    ArrayComparator,
-    cargoErrorMessage,
-    EachValidatorRule,
-    HashAlgorithm,
-    IsUrlOptions,
-    TypedPropertyDecorator,
-    UuidVersion,
-    ValidatorRule,
-} from './types'
-import { CargoClassMetadata } from './metadata'
-import { isDeepEqual } from './utils'
-import { isValidPhoneNumber, CountryCode } from 'libphonenumber-js'
-
-function addValidator(target: any, propertyKey: string | symbol, rule: ValidatorRule, applied?: AppliedDecorator) {
-    const classMeta = new CargoClassMetadata(target)
-    const fieldMeta = classMeta.getFieldMetadata(propertyKey)
-    fieldMeta.addValidator(rule)
-    if (applied) fieldMeta.pushAppliedDecorator(applied)
-    classMeta.setFieldMetadata(propertyKey, fieldMeta)
-}
-
-/**
- * Checks if the value is greater than or equal to the allowed minimum.
- * @param minimum - The minimum allowed value.
- * @param message - Optional custom error message.
- */
-export function Min(minimum: number, message?: cargoErrorMessage): TypedPropertyDecorator<number> {
-    return (target, propertyKey) => {
-        addValidator(
-            target,
-            propertyKey,
-            new ValidatorRule(
-                propertyKey,
-                'min',
-                input => typeof input === 'number' && input >= minimum,
-                message || `${String(propertyKey)} must be >= ${minimum}`,
-            ),
-        )
-    }
-}
-
-/**
- * Checks if the value is less than or equal to the allowed maximum.
- * @param maximum - The maximum allowed value.
- * @param message - Optional custom error message.
- */
-export function Max(maximum: number, message?: cargoErrorMessage): TypedPropertyDecorator<number> {
-    return (target, propertyKey) => {
-        addValidator(
-            target,
-            propertyKey,
-            new ValidatorRule(
-                propertyKey,
-                'max',
-                input => typeof input === 'number' && input <= maximum,
-                message || `${String(propertyKey)} must be <= ${maximum}`,
-            ),
-        )
-    }
-}
-
-/**
- * Checks if the string contains the specified substring.
- * @param seed - The substring to search for.
- * @param message - Optional custom error message.
- */
-export function Contains(seed: string, message?: cargoErrorMessage): TypedPropertyDecorator<string> {
-    return (target, propertyKey) => {
-        addValidator(
-            target,
-            propertyKey,
-            new ValidatorRule(
-                propertyKey,
-                'contains',
-                val => typeof val === 'string' && val.includes(seed),
-                message || `${String(propertyKey)} must contain ${seed}`,
-            ),
-        )
-    }
-}
-
-/**
- * Checks if the string starts with the specified prefix.
- * @param prefixText - The prefix string.
- * @param message - Optional custom error message.
- */
-export function Prefix(prefixText: string, message?: cargoErrorMessage): TypedPropertyDecorator<string> {
-    return (target, propertyKey) => {
-        addValidator(
-            target,
-            propertyKey,
-            new ValidatorRule(
-                propertyKey,
-                'prefix',
-                val => typeof val === 'string' && val.startsWith(prefixText),
-                message || `${String(propertyKey)} must start with ${prefixText}`,
-            ),
-        )
-    }
-}
-
-/**
- * Checks if the string ends with the specified suffix.
- * @param suffixText - The suffix string.
- * @param message - Optional custom error message.
- */
-export function Suffix(suffixText: string, message?: cargoErrorMessage): TypedPropertyDecorator<string> {
-    return (target, propertyKey) => {
-        addValidator(
-            target,
-            propertyKey,
-            new ValidatorRule(
-                propertyKey,
-                'suffix',
-                val => typeof val === 'string' && val.endsWith(suffixText),
-                message || `${String(propertyKey)} must end with ${suffixText}`,
-            ),
-        )
-    }
-}
-
-/**
- * Checks if the value is strictly equal to the specified value.
- * @param value - The value to compare.
- * @param message - Optional custom error message.
- */
-export function Equal(value: any, message?: cargoErrorMessage): PropertyDecorator {
-    return (target: Object, propertyKey: string | symbol): void => {
-        addValidator(
-            target,
-            propertyKey,
-            new ValidatorRule(propertyKey, 'equal', val => val === value, message || `${String(propertyKey)} must be equal to ${value}`),
-        )
-    }
-}
-
-/**
- * Checks if the value is not strictly equal to the specified value.
- * @param value - The value to compare.
- * @param message - Optional custom error message.
- */
-export function NotEqual(value: any, message?: cargoErrorMessage): PropertyDecorator {
-    return (target: Object, propertyKey: string | symbol): void => {
-        addValidator(
-            target,
-            propertyKey,
-            new ValidatorRule(propertyKey, 'notEqual', val => val !== value, message || `${String(propertyKey)} must not be equal to ${value}`),
-        )
-    }
-}
-
-/**
- * Checks if the value is within the specified range (inclusive).
- * @param min - The minimum value.
- * @param max - The maximum value.
- * @param message - Optional custom error message.
- */
-export function Range(min: number, max: number, message?: cargoErrorMessage): TypedPropertyDecorator<number> {
-    return (target, propertyKey) => {
-        addValidator(
-            target,
-            propertyKey,
-            new ValidatorRule(
-                propertyKey,
-                'range',
-                value => typeof value === 'number' && value >= min && value <= max,
-                message || `${String(propertyKey)} must be between ${min} and ${max}`,
-            ),
-        )
-    }
-}
-
-/**
- * Checks if the value is false.
- * @param message - Optional custom error message.
- */
-export function IsFalse(message?: cargoErrorMessage): TypedPropertyDecorator<boolean> {
-    return (target, propertyKey): void => {
-        addValidator(
-            target,
-            propertyKey,
-            new ValidatorRule(propertyKey, 'isFalse', val => val === false, message || `${String(propertyKey)} must be false`),
-        )
-    }
-}
-
-/**
- * Checks if the value is true.
- * @param message - Optional custom error message.
- */
-export function IsTrue(message?: cargoErrorMessage): TypedPropertyDecorator<boolean> {
-    return (target, propertyKey): void => {
-        addValidator(
-            target,
-            propertyKey,
-            new ValidatorRule(propertyKey, 'isTrue', val => val === true, message || `${String(propertyKey)} must be true`),
-        )
-    }
-}
-
-/**
- * Checks if the string length matches the specified value.
- * @param value - The exact length required.
- * @param message - Optional custom error message.
- */
-export function Length(value: number, message?: cargoErrorMessage): TypedPropertyDecorator<string> {
-    return (target, propertyKey): void => {
-        addValidator(
-            target,
-            propertyKey,
-            new ValidatorRule(
-                propertyKey,
-                'length',
-                (val: any) => typeof val === 'string' && val.length === value,
-                message || `${String(propertyKey)} must be ${value} characters`,
-            ),
-        )
-    }
-}
-
-/**
- * Checks if the string length is less than or equal to the maximum.
- * @param max - The maximum length allowed.
- * @param message - Optional custom error message.
- */
-export function MaxLength(max: number, message?: cargoErrorMessage): TypedPropertyDecorator<string> {
-    return (target, propertyKey): void => {
-        addValidator(
-            target,
-            propertyKey,
-            new ValidatorRule(
-                propertyKey,
-                'maxLength',
-                (val: any) => typeof val === 'string' && val.length <= max,
-                message || `${String(propertyKey)} must not exceed ${max} characters`,
-            ),
-        )
-    }
-}
-
-/**
- * Checks if the string length is greater than or equal to the minimum.
- * @param min - The minimum length allowed.
- * @param message - Optional custom error message.
- */
-export function MinLength(min: number, message?: cargoErrorMessage): TypedPropertyDecorator<string> {
-    return (target, propertyKey): void => {
-        addValidator(
-            target,
-            propertyKey,
-            new ValidatorRule(
-                propertyKey,
-                'minLength',
-                (val: any) => typeof val === 'string' && val.length >= min,
-                message || `${String(propertyKey)} must be at least ${min} characters`,
-            ),
-        )
-    }
-}
-
-/**
- * Checks if the value is one of the allowed values.
- * @param options - Array of allowed values.
- * @param message - Optional custom error message.
- */
-export function OneOf<T extends readonly any[]>(options: T, message?: cargoErrorMessage): PropertyDecorator {
-    return (target: Object, propertyKey: string | symbol): void => {
-        addValidator(
-            target,
-            propertyKey,
-            new ValidatorRule(
-                propertyKey,
-                'oneOf',
-                (value: unknown): value is T[number] => options.includes(value as T[number]),
-                message || `${String(propertyKey)} must be one of ${options.join(', ')}`,
-            ),
-        )
-    }
-}
-
-/**
- * Validates the value using a custom validation function.
- * @param validateFn - A function that returns true if valid, false otherwise.
- * @param message - Optional custom error message.
- */
-export function Validate(validateFn: (value: unknown) => boolean, message?: cargoErrorMessage): PropertyDecorator {
-    return (target: Object, propertyKey: string | symbol): void => {
-        addValidator(
-            target,
-            propertyKey,
-            new ValidatorRule(propertyKey, 'validate', validateFn, message || `${String(propertyKey)} did not pass the provided validation rule.`),
-        )
-    }
-}
-
-/**
- * Checks if the string matches the specified regular expression.
- * @param pattern - The regular expression to match against.
- * @param message - Optional custom error message.
- */
-export function Regexp(pattern: RegExp, message?: cargoErrorMessage): TypedPropertyDecorator<string> {
-    return (target, propertyKey): void => {
-        addValidator(
-            target,
-            propertyKey,
-            new ValidatorRule(
-                propertyKey,
-                'regexp',
-                (value: unknown) => typeof value === 'string' && pattern.test(value),
-                message || `${String(propertyKey)} does not match pattern ${pattern}`,
-            ),
-        )
-    }
-}
-
-/**
- * Checks if the string is a valid email address.
- * @param message - Optional custom error message.
- */
-export function Email(message?: cargoErrorMessage): TypedPropertyDecorator<string> {
-    return (target, propertyKey): void => {
-        const DEFAULT_EMAIL_PATTERN =
-            /^[a-zA-Z0-9.!#$%&'*+/=?^_`{|}~-]+@[a-zA-Z0-9](?:[a-zA-Z0-9-]{0,61}[a-zA-Z0-9])?(?:\.[a-zA-Z0-9](?:[a-zA-Z0-9-]{0,61}[a-zA-Z0-9])?)*$/
-        addValidator(
-            target,
-            propertyKey,
-            new ValidatorRule(
-                propertyKey,
-                'email',
-                (value: unknown) => typeof value === 'string' && DEFAULT_EMAIL_PATTERN.test(value),
-                message || `${String(propertyKey)} should be email format`,
-            ),
-        )
-    }
-}
-
-const ALPHA_PATTERN = /^[a-zA-Z]+$/
-
-/**
- * Checks if the string contains only alphabetic characters.
- * @param message - Optional custom error message.
- */
-export function Alpha(message?: cargoErrorMessage): TypedPropertyDecorator<string> {
-    return (target, propertyKey): void => {
-        addValidator(
-            target,
-            propertyKey,
-            new ValidatorRule(
-                propertyKey,
-                'alpha',
-                (value: unknown) => typeof value === 'string' && ALPHA_PATTERN.test(value),
-                message || `${String(propertyKey)} should be alphabetic`,
-            ),
-        )
-    }
-}
-
-const uuidPatterns = {
-    all: /^[0-9a-f]{8}-[0-9a-f]{4}-[1345][0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$/i,
-    v1: /^[0-9a-f]{8}-[0-9a-f]{4}-1[0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$/i,
-    v3: /^[0-9a-f]{8}-[0-9a-f]{4}-3[0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$/i,
-    v4: /^[0-9a-f]{8}-[0-9a-f]{4}-4[0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$/i,
-    v5: /^[0-9a-f]{8}-[0-9a-f]{4}-5[0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$/i,
-}
-
-/**
- * Checks if the string is a valid UUID.
- * @param version - The UUID version to check against ('v1', 'v3', 'v4', 'v5', or 'all').
- * @param message - Optional custom error message.
- */
-export function Uuid(version?: UuidVersion, message?: cargoErrorMessage): TypedPropertyDecorator<string> {
-    let regex: RegExp
-    let versionLabel: string
-
-    const v = version ? version.replace('v', '') : 'all'
-
-    switch (v) {
-        case '1':
-            regex = uuidPatterns.v1
-            versionLabel = 'v1'
-            break
-        case '3':
-            regex = uuidPatterns.v3
-            versionLabel = 'v3'
-            break
-        case '4':
-            regex = uuidPatterns.v4
-            versionLabel = 'v4'
-            break
-        case '5':
-            regex = uuidPatterns.v5
-            versionLabel = 'v5'
-            break
-        default:
-            regex = uuidPatterns.all
-            versionLabel = 'v1, v3, v4, or v5'
-    }
-
-    return (target, propertyKey): void => {
-        addValidator(
-            target,
-            propertyKey,
-            new ValidatorRule(
-                propertyKey,
-                'uuid',
-                (value: unknown) => typeof value === 'string' && regex.test(value),
-                message || `${String(propertyKey)} must be a valid UUID format (${versionLabel})`,
-            ),
-        )
-    }
-}
-
-/**
- * Checks if the string contains only alphanumeric characters.
- * @param message - Optional custom error message.
- */
-export function Alphanumeric(message?: cargoErrorMessage): TypedPropertyDecorator<string> {
-    return (target, propertyKey): void => {
-        addValidator(
-            target,
-            propertyKey,
-            new ValidatorRule(
-                propertyKey,
-                'alphanumeric',
-                (value: unknown) => typeof value === 'string' && /^[a-zA-Z0-9]+$/.test(value),
-                message || `${String(propertyKey)} should be alphanumeric`,
-            ),
-        )
-    }
-}
-
-/**
- * Checks if the string contains only uppercase characters.
- * @param message - Optional custom error message.
- */
-export function IsUppercase(message?: cargoErrorMessage): TypedPropertyDecorator<string> {
-    return (target, propertyKey): void => {
-        addValidator(
-            target,
-            propertyKey,
-            new ValidatorRule(
-                propertyKey,
-                'isUppercase',
-                (value: unknown) => typeof value === 'string' && value.toUpperCase() === value,
-                message || `${String(propertyKey)} should be uppercase`,
-            ),
-        )
-    }
-}
-
-/**
- * Checks if the string is a valid JSON Web Token (JWT).
- * @param message - Optional custom error message.
- */
-const JWT_PATTERN = /^[A-Za-z0-9\-_]+\.[A-Za-z0-9\-_]+\.[A-Za-z0-9\-_]*$/
-
-export function IsJwt(message?: cargoErrorMessage): TypedPropertyDecorator<string> {
-    return (target, propertyKey): void => {
-        addValidator(
-            target,
-            propertyKey,
-            new ValidatorRule(
-                propertyKey,
-                'isJwt',
-                (value: unknown) => typeof value === 'string' && JWT_PATTERN.test(value),
-                message || `${String(propertyKey)} should be a valid JWT`,
-            ),
-        )
-    }
-}
-
-/**
- * Checks if the string is a valid phone number for the given region.
- * If no region is provided, validates as an international number (must include country code).
- * @param region - Optional ISO 3166-1 alpha-2 region code (e.g., 'KR', 'US').
- * @param message - Optional custom error message.
- */
-export function IsPhoneNumber(region?: CountryCode, message?: cargoErrorMessage): TypedPropertyDecorator<string> {
-    return (target, propertyKey): void => {
-        addValidator(
-            target,
-            propertyKey,
-            new ValidatorRule(
-                propertyKey,
-                'isPhoneNumber',
-                (value: unknown) => typeof value === 'string' && isValidPhoneNumber(value, region),
-                message || `${String(propertyKey)} must be a valid phone number${region ? ` for ${region}` : ''}`,
-            ),
-        )
-    }
-}
-
-/**
- * Checks if the string is a valid IANA timezone identifier.
- * @param message - Optional custom error message.
- */
-export function IsTimeZone(message?: cargoErrorMessage): TypedPropertyDecorator<string> {
-    return (target, propertyKey): void => {
-        addValidator(
-            target,
-            propertyKey,
-            new ValidatorRule(
-                propertyKey,
-                'isTimeZone',
-                (value: unknown) => {
-                    if (typeof value !== 'string') return false
-                    try {
-                        Intl.DateTimeFormat(undefined, { timeZone: value })
-                        return true
-                    } catch {
-                        return false
-                    }
-                },
-                message || `${String(propertyKey)} must be a valid IANA timezone`,
-            ),
-        )
-    }
-}
-
-/**
- * Checks if the string is a valid hex color code.
- * Supports #RGB, #RGBA, #RRGGBB, #RRGGBBAA formats.
- * @param message - Optional custom error message.
- */
-const HEX_COLOR_PATTERN = /^#([0-9a-f]{3}|[0-9a-f]{4}|[0-9a-f]{6}|[0-9a-f]{8})$/i
-
-export function IsHexColor(message?: cargoErrorMessage): TypedPropertyDecorator<string> {
-    return (target, propertyKey): void => {
-        addValidator(
-            target,
-            propertyKey,
-            new ValidatorRule(
-                propertyKey,
-                'isHexColor',
-                (value: unknown) => typeof value === 'string' && HEX_COLOR_PATTERN.test(value),
-                message || `${String(propertyKey)} should be a hex color code`,
-            ),
-        )
-    }
-}
-
-/**
- * Checks if the string is a valid hexadecimal number.
- * @param message - Optional custom error message.
- */
-const HEX_PATTERN = /^(0x)?[0-9a-fA-F]+$/i
-
-export function IsHexadecimal(message?: cargoErrorMessage): TypedPropertyDecorator<string> {
-    return (target, propertyKey): void => {
-        addValidator(
-            target,
-            propertyKey,
-            new ValidatorRule(
-                propertyKey,
-                'isHexadecimal',
-                (value: unknown) => typeof value === 'string' && HEX_PATTERN.test(value),
-                message || `${String(propertyKey)} should be a hexadecimal number`,
-            ),
-        )
-    }
-}
-
-const hashPatterns: Record<HashAlgorithm, RegExp> = {
-    md5: /^[0-9a-f]{32}$/i,
-    sha1: /^[0-9a-f]{40}$/i,
-    sha256: /^[0-9a-f]{64}$/i,
-    sha384: /^[0-9a-f]{96}$/i,
-    sha512: /^[0-9a-f]{128}$/i,
-    crc32: /^[0-9a-f]{8}$/i,
-    crc32b: /^[0-9a-f]{8}$/i,
-}
-
-/**
- * Checks if the string is a valid hash for the specified algorithm.
- * Supported algorithms: md5, sha1, sha256, sha384, sha512, crc32, crc32b.
- * @param algorithm - The hash algorithm to validate against.
- * @param message - Optional custom error message.
- */
-export function IsHash(algorithm: HashAlgorithm, message?: cargoErrorMessage): TypedPropertyDecorator<string> {
-    const pattern = hashPatterns[algorithm]
-    return (target, propertyKey): void => {
-        addValidator(
-            target,
-            propertyKey,
-            new ValidatorRule(
-                propertyKey,
-                'isHash',
-                (value: unknown) => typeof value === 'string' && pattern.test(value),
-                message || `${String(propertyKey)} should be a valid ${algorithm} hash`,
-            ),
-        )
-    }
-}
-
-/**
- * Checks if the Date value is greater than or equal to the given minimum date.
- * @param min - The minimum allowed date, or a function that returns it.
- * @param message - Optional custom error message.
- */
-export function MinDate(min: Date | (() => Date), message?: cargoErrorMessage): TypedPropertyDecorator<Date> {
-    return (target, propertyKey): void => {
-        let minDate: Date
-        addValidator(
-            target,
-            propertyKey,
-            new ValidatorRule(
-                propertyKey,
-                'minDate',
-                (value: unknown) => {
-                    minDate = typeof min === 'function' ? min() : min
-                    return value instanceof Date && !isNaN(value.getTime()) && value >= minDate
-                },
-                message || (() => `${String(propertyKey)} must be after ${minDate}`),
-            ),
-        )
-    }
-}
-
-/**
- * Checks if the Date value is less than or equal to the given maximum date.
- * @param max - The maximum allowed date, or a function that returns it.
- * @param message - Optional custom error message.
- */
-export function MaxDate(max: Date | (() => Date), message?: cargoErrorMessage): TypedPropertyDecorator<Date> {
-    return (target, propertyKey): void => {
-        let maxDate: Date
-        addValidator(
-            target,
-            propertyKey,
-            new ValidatorRule(
-                propertyKey,
-                'maxDate',
-                (value: unknown) => {
-                    maxDate = typeof max === 'function' ? max() : max
-                    return value instanceof Date && !isNaN(value.getTime()) && value <= maxDate
-                },
-                message || (() => `${String(propertyKey)} must be before ${maxDate}`),
-            ),
-        )
-    }
-}
-
-/**
- * Checks if the string is a valid URL.
- * @param options - Optional configuration (e.g., allowed protocols).
- * @param message - Optional custom error message.
- */
-export function IsUrl(options?: IsUrlOptions, message?: cargoErrorMessage): TypedPropertyDecorator<string> {
-    const protocols = options?.protocols ?? ['http', 'https', 'ftp']
-    return (target, propertyKey): void => {
-        addValidator(
-            target,
-            propertyKey,
-            new ValidatorRule(
-                propertyKey,
-                'isUrl',
-                (value: unknown) => {
-                    if (typeof value !== 'string') return false
-                    try {
-                        const url = new URL(value)
-                        const protocol = url.protocol.replace(/:$/, '')
-                        return protocols.includes(protocol)
-                    } catch {
-                        return false
-                    }
-                },
-                message || `${String(propertyKey)} should be a valid URL`,
-            ),
-        )
-    }
-}
-
-/**
- * Checks if the string contains only lowercase characters.
- * @param message - Optional custom error message.
- */
-export function IsLowercase(message?: cargoErrorMessage): TypedPropertyDecorator<string> {
-    return (target, propertyKey): void => {
-        addValidator(
-            target,
-            propertyKey,
-            new ValidatorRule(
-                propertyKey,
-                'isLowercase',
-                (value: unknown) => typeof value === 'string' && value.toLowerCase() === value,
-                message || `${String(propertyKey)} should be lowercase`,
-            ),
-        )
-    }
-}
-
-/**
- * Validates that if the decorated property has a value, the specified field must also be present.
- * @param fieldName - The name of the required field.
- * @param message - Optional custom error message.
- */
-export function With(fieldName: string, message?: cargoErrorMessage): PropertyDecorator {
-    return (target, propertyKey): void => {
-        addValidator(
-            target,
-            propertyKey,
-            new ValidatorRule(
-                propertyKey,
-                'with',
-                (value: unknown, instance?: Record<string | symbol, any>) => !(!!value && !instance?.[fieldName]),
-                message || `${String(propertyKey)} requires ${fieldName}`,
-            ),
-            { name: With.name, category: 'validator', args: [fieldName] },
-        )
-    }
-}
-
-/**
- * Validates that if the decorated property has a value, the specified field must NOT be present.
- * @param fieldName - The name of the field that must be absent.
- * @param message - Optional custom error message.
- */
-export function Without(fieldName: string, message?: cargoErrorMessage): PropertyDecorator {
-    return (target, propertyKey): void => {
-        addValidator(
-            target,
-            propertyKey,
-            new ValidatorRule(
-                propertyKey,
-                'without',
-                (value: unknown, instance?: Record<string | symbol, any>) => {
-                    if (!instance) return false
-                    return !(!!value && !!instance?.[fieldName])
-                },
-                message || `${String(propertyKey)} cannot exist with ${fieldName}`,
-            ),
-            { name: Without.name, category: 'validator', args: [fieldName] },
-        )
-    }
-}
-
-/**
- * Checks if the array contains all the specified values.
- * @param values - The values that must be present in the array.
- * @param comparator - Optional custom comparison function (expected, actual) => boolean.
- *                     When provided, all comparisons are delegated to this function,
- *                     including primitives.
- * @param message - Optional custom error message.
- */
-export function ListContains(values: any[], comparator?: ArrayComparator, message?: cargoErrorMessage): TypedPropertyDecorator<any[]> {
-    // Pre-split only when using default comparison (Set + deepEqual optimization)
-    const expectedPrimitives = !comparator ? values.filter(v => v === null || typeof v !== 'object') : []
-    const expectedObjects = !comparator ? values.filter(v => v !== null && typeof v === 'object') : []
-
-    return (target: Object, propertyKey: string | symbol): void => {
-        addValidator(
-            target,
-            propertyKey,
-            new ValidatorRule(
-                propertyKey,
-                'listContains',
-                (value: unknown) => {
-                    if (!Array.isArray(value)) {
-                        return false
-                    }
-
-                    // Delegate all comparisons to the custom comparator, including primitives
-                    if (comparator) {
-                        return values.every(expected => value.some(actual => comparator(expected, actual)))
-                    }
-
-                    const actualPrimitiveSet = new Set()
-                    const actualObjects: any[] = []
-
-                    for (const item of value) {
-                        if (item === null || typeof item !== 'object') {
-                            actualPrimitiveSet.add(item)
-                        } else {
-                            actualObjects.push(item)
-                        }
-                    }
-
-                    // Verify all expected primitive values exist in the actual array
-                    for (const req of expectedPrimitives) {
-                        if (!actualPrimitiveSet.has(req)) return false
-                    }
-
-                    // Verify all expected objects exist in the actual array using deep equality
-                    for (const reqObj of expectedObjects) {
-                        const found = actualObjects.some(actObj => isDeepEqual(reqObj, actObj))
-                        if (!found) return false
-                    }
-
-                    return true
-                },
-                message || `${String(propertyKey)} must contain all specified values`,
-            ),
-        )
-    }
-}
-
-/**
- * Validates that the array does NOT contain any of the specified values.
- * Supports primitives, objects (deep equality), Date values, and custom comparators.
- * @param values - The values that must NOT be present in the array.
- * @param comparator - Optional custom comparison function for comparing values,
- *                     including primitives.
- * @param message - Optional custom error message.
- */
-export function ListNotContains(values: any[], comparator?: ArrayComparator, message?: cargoErrorMessage): TypedPropertyDecorator<any[]> {
-    // Pre-split only when using default comparison (Set + deepEqual optimization)
-    const excludedPrimitives = !comparator ? values.filter(v => v === null || typeof v !== 'object') : []
-    const excludedObjects = !comparator ? values.filter(v => v !== null && typeof v === 'object') : []
-
-    return (target: Object, propertyKey: string | symbol): void => {
-        addValidator(
-            target,
-            propertyKey,
-            new ValidatorRule(
-                propertyKey,
-                'listNotContains',
-                (value: unknown) => {
-                    if (!Array.isArray(value)) {
-                        return false
-                    }
-
-                    // Delegate all comparisons to the custom comparator, including primitives
-                    if (comparator) {
-                        return values.every(excluded => !value.some(actual => comparator(excluded, actual)))
-                    }
-
-                    const actualPrimitiveSet = new Set()
-                    const actualObjects: any[] = []
-
-                    for (const item of value) {
-                        if (item === null || typeof item !== 'object') {
-                            actualPrimitiveSet.add(item)
-                        } else {
-                            actualObjects.push(item)
-                        }
-                    }
-
-                    // Verify none of the excluded primitive values exist in the actual array
-                    for (const exc of excludedPrimitives) {
-                        if (actualPrimitiveSet.has(exc)) return false
-                    }
-
-                    // Verify none of the excluded objects exist in the actual array using deep equality
-                    for (const excObj of excludedObjects) {
-                        const found = actualObjects.some(actObj => isDeepEqual(excObj, actObj))
-                        if (found) return false
-                    }
-
-                    return true
-                },
-                message || `${String(propertyKey)} must not contain any of the excluded values`,
-            ),
-        )
-    }
-}
-
-/**
- * Checks if the array length does not exceed the specified maximum size.
- * @param max - The maximum number of elements allowed in the array.
- * @param message - Optional custom error message.
- */
-export function ListMaxSize(max: number, message?: cargoErrorMessage): TypedPropertyDecorator<any[]> {
-    return (target: Object, propertyKey: string | symbol): void => {
-        addValidator(
-            target,
-            propertyKey,
-            new ValidatorRule(
-                propertyKey,
-                'listMaxSize',
-                (value: unknown) => Array.isArray(value) && value.length <= max,
-                message || `${String(propertyKey)} must contain no more than ${max} elements`,
-            ),
-        )
-    }
-}
-
-/**
- * Checks if the array length does not fall below the specified minimum size.
- * @param min - The minimum number of elements allowed in the array.
- * @param message - Optional custom error message.
- */
-export function ListMinSize(min: number, message?: cargoErrorMessage): TypedPropertyDecorator<any[]> {
-    return (target: Object, propertyKey: string | symbol): void => {
-        addValidator(
-            target,
-            propertyKey,
-            new ValidatorRule(
-                propertyKey,
-                'listMinSize',
-                (value: unknown) => Array.isArray(value) && value.length >= min,
-                message || `${String(propertyKey)} must contain at least ${min} elements`,
-            ),
-        )
-    }
-}
-
-/**
- * Applies validation rules to each element of an array.
- * @param args - Validation decorators or functions to apply to each element.
- */
-export function Each(...args: (PropertyDecorator | TypedPropertyDecorator<any> | ((value: any) => boolean))[]): PropertyDecorator {
-    return (target: any, propertyKey: string | symbol) => {
-        const classMeta = new CargoClassMetadata(target)
-        const fieldMeta = classMeta.getFieldMetadata(propertyKey)
-
-        args.forEach(arg => {
-            // Create a temporary class to determine if 'arg' is a decorator
-            const tempClass = class {}
-            const tempKey = 'temp'
-            let rulesAdded: ValidatorRule[] = []
-
-            try {
-                // Check the number of validators before applying the decorator
-                const tempMetaBefore = new CargoClassMetadata(tempClass.prototype)
-                const fieldBefore = tempMetaBefore.getFieldMetadata(tempKey)
-                const validatorsBefore = fieldBefore.getValidators().length
-
-                // Attempt to execute as a PropertyDecorator
-                ;(arg as PropertyDecorator)(tempClass.prototype, tempKey)
-
-                // Check if a new validator rule has been added
-                const tempMetaAfter = new CargoClassMetadata(tempClass.prototype)
-                const tempFieldAfter = tempMetaAfter.getFieldMetadata(tempKey)
-                const validatorsAfter = tempFieldAfter.getValidators()
-
-                if (validatorsAfter.length > validatorsBefore) {
-                    rulesAdded = validatorsAfter.slice(validatorsBefore)
-                }
-
-                // Record any decorator tags applied by `arg` so schema-validation rule checkers can inspect what was nested inside `@Each`.
-                tempFieldAfter.getAppliedDecorators().forEach(tag => fieldMeta.pushAppliedDecorator(tag, 'each'))
-            } catch (e) {
-                // Ignore errors if 'arg' is not a valid decorator
-            }
-
-            // Wrap the extracted/created rule with EachValidatorRule
-            if (rulesAdded.length > 0) {
-                rulesAdded.forEach(rule => {
-                    fieldMeta.addValidator(new EachValidatorRule(propertyKey, rule))
-                })
-            } else if (typeof arg === 'function') {
-                // If it's not a decorator but a plain function, treat it as a custom validator
-                const rule = new ValidatorRule(propertyKey, 'custom', arg as (value: any) => boolean, `Validation failed for ${String(propertyKey)}`)
-                fieldMeta.addValidator(new EachValidatorRule(propertyKey, rule))
-            }
-        })
-
-        fieldMeta.pushAppliedDecorator({ name: Each.name, category: 'validator', args })
-        classMeta.setFieldMetadata(propertyKey, fieldMeta)
-    }
-}
