+++ packages/express-cargo/tests/decorator/enum.test.ts
@@ -1,6 +1,6 @@
 import { CargoFieldError } from '../../src'
 import { CargoClassMetadata } from '../../src/metadata'
-import { Enum } from '../../src/enum'
+import { Enum } from '../../src/decorators/typeHelper'
 
 enum Role {
     ADMIN = 'admin',
